mnist-digit-detector/requirements.txt
streamlit
tensorflow
pillow
numpy
matplotlib